TS-92: VIO and VConnection interfaces now use ink64 instead of int for all sizes related to buffers or objects.  Counts remain int. This checkin fixes type warnings in the modules which uses these interfaces.
git-svn-id: https://svn.apache.org/repos/asf/incubator/trafficserver/traffic/branches/dev@897010 13f79535-47bb-0310-9956-ffa450edef68

+++ iocore/cache/RamCache.cc
@@ -247,7 +247,7 @@ RamCache::print_stats(FILE * fp, int verbose)
       for (int j = 0; j < partition_size; j++) {
         RamCacheEntry *e = partition[i].bucket[j].head;
         while (e) {
-          printf("\t\t%9d %9d\n", e->data->_size_index, e->data->block_size());
+          printf("\t\t%9d %9d\n", (int)e->data->_size_index, (int)e->data->block_size());
           total += e->data->block_size();
           n++;
           e = e->hash_link.next;
@@ -260,7 +260,7 @@ RamCache::print_stats(FILE * fp, int verbose)
       {
         RamCacheEntry *e = partition[i].lru.head;
         while (e) {
-          printf("\t\t%9d %9d\n", e->data->_size_index, e->data->block_size());
+          printf("\t\t%9d %9d\n", (int)e->data->_size_index, (int)e->data->block_size());
           total += e->data->block_size();
           n++;
           e = e->lru_link.next;
